requirements.txt
pandas>=2.2.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
torch>=2.0.0
torchvision>=0.15.0
xgboost>=1.7.0
opencv-python>=4.8.0
Pillow>=10.0.0
requests>=2.31.0
openpyxl>=3.1.0
geopandas>=0.13.0
folium>=0.14.0
shap>=0.42.0
pyyaml>=6.0.0
jupyter>=1.0.0
ipykernel>=6.25.0
python-dotenv>=1.0.0
tqdm>=4.66.0
